Additional optimizations.
Erlang is more efficient making a single request per light thread
than batching. When batching more than 1, the context switch is
800ms per thread. When reducing the batch size to 1 the context
switch drops an order of magnatude to 80ms, which is basically the
RPC request + processing the request. If the go-pulse servers were
not local to us and we had rate-limits, it would make sense to batch
but we control everthing and the nodes are local, so it's cheaper to
make 10,000 requests per second both on the go-pulse side and blockscout.